18-08-2024_1 Alpha: GLTFLoader and Opengl ES framebuffer fix, Application is now a proper singleton, removed config.h and more

+++ src/Tabby/Renderer/Texture.h
@@ -1,10 +1,6 @@
 #pragma once
-
-#include "Tabby/Core/Base.h"
-#include <Tabby/Asset/AssetBase.h>
-
 #include <Tabby/Foundation/Types.h>
-#include <string>
+#include <Tabby/Asset/AssetBase.h>
 
 namespace Tabby {
 
@@ -25,25 +21,6 @@ enum class ImageUsage : uint8_t {
     RENDER_TARGET,
     DEPTH_BUFFER
 };
-// enum class ImageFormat : uint8_t {
-//     R8,
-//     RB16,
-//     RGB24,
-//     RGBA32_SRGB,
-//     RGBA32_UNORM,
-//     BGRA32_SRGB,
-//     BGRA32_UNORM,
-//     RGB32_HDR,
-//     RGBA64_HDR,
-//     RGBA128_HDR,
-//     D32,
-//     BC1,
-//     BC5,
-//     BC6h,
-//     BC7,
-//     RGBA64_SFLOAT,
-//     RGB24_UNORM
-// };
 
 enum class ImageFormat : uint8_t {
     None = 0,
@@ -84,7 +61,7 @@ struct TextureSpecification {
     ImageUsage usage = ImageUsage::TEXTURE;
     ImageType type = ImageType::TYPE_2D;
     uint8_t array_layers = 1;
-    uint32_t UnpackAlignment = 4;
+    uint8_t UnpackAlignment = 4;
     bool GenerateMips = true;
 };
